refactor : modified for moving files to other directory
