+++ src/main/java/com/swifty/bank/server/BankApplication.java
@@ -1,4 +1,4 @@
-package com.swifty.bank;
+package com.swifty.bank.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
