+++ src/main/java/com/swifty/bank/server/core/customer/constant/CustomerStatus.java
@@ -1,4 +1,4 @@
-package com.swifty.bank.server.src.main.core.customer.constant;
+package com.swifty.bank.server.core.customer.constant;
 
 public enum CustomerStatus {
     ACTIVE("나 살아있다"),
