+++ src/main/java/com/swifty/bank/server/core/customer/constant/Nationality.java
@@ -1,4 +1,4 @@
-package com.swifty.bank.server.src.main.core.customer.constant;
+package com.swifty.bank.server.core.customer.constant;
 
 public enum Nationality {
     KOREA("한국");
